+++ docs/pipeline_jerarquico
