+++ src/main.py
@@ -1,0 +1,5 @@
+def main():
+    print("hola")
+
+if __name__ == "__main__":
+    main()
